Build NPCs only if specified

+++ Jenkins/build-simulator.sh
@@ -234,8 +234,10 @@ cp /mnt/AssetBundles/Controllables/controllable_* /tmp/${BUILD_OUTPUT}/AssetBund
 mkdir -p /tmp/${BUILD_OUTPUT}/AssetBundles/Sensors
 cp /mnt/AssetBundles/Sensors/sensor_* /tmp/${BUILD_OUTPUT}/AssetBundles/Sensors
 
-mkdir -p /tmp/${BUILD_OUTPUT}/AssetBundles/NPCs
-cp -R /mnt/AssetBundles/NPCs/* /tmp/${BUILD_OUTPUT}/AssetBundles/NPCs
+if [ ! -z ${SIMULATOR_NPCS+x} ]; then
+    mkdir -p /tmp/${BUILD_OUTPUT}/AssetBundles/NPCs
+    cp -R /mnt/AssetBundles/NPCs/* /tmp/${BUILD_OUTPUT}/AssetBundles/NPCs
+fi
 
 cd /tmp
 zip -r /mnt/${BUILD_OUTPUT}.zip ${BUILD_OUTPUT}
